--- unit_tests/parent_directory.py
@@ -1,5 +1,0 @@
-import os, sys, inspect
-
-parent_dir = os.path.dirname(os.path.dirname(os.path.abspath(inspect.getfile(inspect.currentframe()))))
-
-sys.path.insert(0, parent_dir)
