+++ unit_tests/test_file_io.py
@@ -1,4 +1,4 @@
-import parent_directory
+import simple_imports
 
 import os
 import unittest
