+++ unit_tests/test_splitter.py
@@ -1,4 +1,4 @@
-import parent_directory
+import simple_imports
 
 import unittest
 from data_splitter import DataSplitter
